HEFP encoding improvements
